Разбил gdb_wrapper на отдельные файлы
